Nifi-12193 Implement boolean switch for python processor support
