+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-resources/src/main/resources/conf/nifi.properties
@@ -44,7 +44,8 @@ nifi.nar.unpack.uber.jar=${nifi.nar.unpack.uber.jar}
 # Python Extensions #
 #####################
 # Uncomment in order to enable Python Extensions.
-#nifi.python.command=${nifi.python.command}
+nifi.python.enabled=${nifi.python.enabled}
+nifi.python.command=${nifi.python.command}
 nifi.python.framework.source.directory=${nifi.python.framework.source.directory}
 nifi.python.extensions.source.directory.default=${nifi.python.extensions.source.directory.default}
 nifi.python.working.directory=${nifi.python.working.directory}
