+++ nifi-system-tests/nifi-system-test-suite/src/test/resources/conf/pythonic/nifi.properties
@@ -46,6 +46,7 @@ nifi.documentation.working.directory=./work/docs/components
 # Python Extensions #
 #####################
 # Uncomment in order to enable Python Extensions.
+nifi.python.enabled=true
 nifi.python.command=python3
 nifi.python.framework.source.directory=./python/framework
 nifi.python.extensions.source.directory.default=./python/extensions
